New version requires setting config source.
Previous version did not allow for breaking changes. New version
allows the default source to change over time because users always
have to set it.

If users don't set a source for config, we use a null deserializer
and fall back to the default impl for the config, never attempting
to load config from the environment.

Docs need to change.